backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.1
